requirements.txt
# PyTorch and CUDA Support (commented out - enable if needed)
# torch
# torchvision
# torchaudio

# Transformers and Embedding Models
transformers>=4.41.0,<5.0.0
sentence-transformers==3.3.1

# Web Framework & API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart==0.0.19
starlette>=0.36.0
aiofiles>=23.2.0
websockets>=12.0

# Database & Vector Store
chromadb==0.4.24
langchain-core>=0.1.15
langchain-community>=0.0.20
langchain-chroma>=0.0.5

# PostgreSQL Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# LLM Integrations
langchain-ollama>=0.0.3
langchain-google-genai>=0.0.10
google-generativeai>=0.3.2

# File Processing
python-pptx>=0.6.21
PyPDF2>=3.0.0
nbformat>=5.9.2
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Optional: Development Tools
pytest>=7.4.0  # For testing
black>=23.12.0  # For code formatting

# Hume AI - Latest Version
hume>=0.8.6

# Additional dependencies for async operations
httpx>=0.27.0
python-jose>=3.3.0  # For JWT token handling

PyJWT>=2.0.0